+++ dysmsapi-20180501/alibabacloud_dysmsapi20180501/client.py
@@ -55,10 +55,12 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.BatchSendMessageToGlobeResponse:
         """
-        You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
-        *   You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
-        *   For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
-        *   In each request, you can send messages to up to 1,000 mobile phone numbers.
+        @summary Sends a message to multiple recipients in countries or regions outside the Chinese mainland.
+        
+        @description    You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
+        You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
+        For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
+        In each request, you can send messages to up to 1,000 mobile phone numbers.
         ### [](#qps)QPS limits
         You can call this operation only once per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
         
@@ -105,10 +107,12 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.BatchSendMessageToGlobeResponse:
         """
-        You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
-        *   You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
-        *   For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
-        *   In each request, you can send messages to up to 1,000 mobile phone numbers.
+        @summary Sends a message to multiple recipients in countries or regions outside the Chinese mainland.
+        
+        @description    You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
+        You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
+        For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
+        In each request, you can send messages to up to 1,000 mobile phone numbers.
         ### [](#qps)QPS limits
         You can call this operation only once per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
         
@@ -154,10 +158,12 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.BatchSendMessageToGlobeRequest,
     ) -> dysmsapi_20180501_models.BatchSendMessageToGlobeResponse:
         """
-        You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
-        *   You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
-        *   For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
-        *   In each request, you can send messages to up to 1,000 mobile phone numbers.
+        @summary Sends a message to multiple recipients in countries or regions outside the Chinese mainland.
+        
+        @description    You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
+        You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
+        For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
+        In each request, you can send messages to up to 1,000 mobile phone numbers.
         ### [](#qps)QPS limits
         You can call this operation only once per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
         
@@ -172,10 +178,12 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.BatchSendMessageToGlobeRequest,
     ) -> dysmsapi_20180501_models.BatchSendMessageToGlobeResponse:
         """
-        You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
-        *   You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
-        *   For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
-        *   In each request, you can send messages to up to 1,000 mobile phone numbers.
+        @summary Sends a message to multiple recipients in countries or regions outside the Chinese mainland.
+        
+        @description    You cannot call the BatchSendMessageToGlobe operation to send messages to the Chinese mainland.
+        You can call the BatchSendMessageToGlobe operation to send notifications and promotional messages to a limited number of mobile phone numbers at a time. To send messages to a large number of mobile phone numbers at a time, use the mass messaging feature of the SMS console.
+        For time-sensitive related messages, we recommend that you use the [SendMessageToGlobe](https://www.alibabacloud.com/help/zh/sms/developer-reference/api-dysmsapi-2018-05-01-batchsendmessagetoglobe) operation to ensure that messages are delivered on time.
+        In each request, you can send messages to up to 1,000 mobile phone numbers.
         ### [](#qps)QPS limits
         You can call this operation only once per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
         
@@ -191,9 +199,11 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.ConversionDataResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary This API, sends conversion data to the Alibaba SMS service.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         
         @param request: ConversionDataRequest
@@ -231,9 +241,11 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.ConversionDataResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary This API, sends conversion data to the Alibaba SMS service.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         
         @param request: ConversionDataRequest
@@ -270,9 +282,11 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.ConversionDataRequest,
     ) -> dysmsapi_20180501_models.ConversionDataResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary This API, sends conversion data to the Alibaba SMS service.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         
         @param request: ConversionDataRequest
@@ -286,9 +300,11 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.ConversionDataRequest,
     ) -> dysmsapi_20180501_models.ConversionDataResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary This API, sends conversion data to the Alibaba SMS service.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         
         @param request: ConversionDataRequest
@@ -303,7 +319,9 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.QueryMessageResponse:
         """
-        ### QPS limit
+        @summary Queries the delivery report of a message.
+        
+        @description ### QPS limit
         You can call this operation up to 300 times per second. If the number of the calls per second exceeds a limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limits when you call this operation.
         
         @param request: QueryMessageRequest
@@ -339,7 +357,9 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.QueryMessageResponse:
         """
-        ### QPS limit
+        @summary Queries the delivery report of a message.
+        
+        @description ### QPS limit
         You can call this operation up to 300 times per second. If the number of the calls per second exceeds a limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limits when you call this operation.
         
         @param request: QueryMessageRequest
@@ -374,7 +394,9 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.QueryMessageRequest,
     ) -> dysmsapi_20180501_models.QueryMessageResponse:
         """
-        ### QPS limit
+        @summary Queries the delivery report of a message.
+        
+        @description ### QPS limit
         You can call this operation up to 300 times per second. If the number of the calls per second exceeds a limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limits when you call this operation.
         
         @param request: QueryMessageRequest
@@ -388,7 +410,9 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.QueryMessageRequest,
     ) -> dysmsapi_20180501_models.QueryMessageResponse:
         """
-        ### QPS limit
+        @summary Queries the delivery report of a message.
+        
+        @description ### QPS limit
         You can call this operation up to 300 times per second. If the number of the calls per second exceeds a limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limits when you call this operation.
         
         @param request: QueryMessageRequest
@@ -403,10 +427,12 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SendMessageToGlobeResponse:
         """
-        ### Usage notes
-        You cannot call the SendMessageToGlobe operation to send messages to the Chinese mainland.
-        ### QPS limit
-        You may call this operation up to 300 times per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
+        @summary Sends a message to regions outside the Chinese mainland.
+        
+        @description ### [](#)Usage notes
+        The SendMessageToGlobe API operation does not support message delivery to the Chinese mainland.
+        ### [](#qps-)QPS limit
+        You can call this operation up to 30 times per second per account. If the number of calls per second exceeds the limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limit when you call this operation.
         
         @param request: SendMessageToGlobeRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -414,6 +440,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.channel_id):
+            query['ChannelId'] = request.channel_id
         if not UtilClient.is_unset(request.from_):
             query['From'] = request.from_
         if not UtilClient.is_unset(request.message):
@@ -449,10 +477,12 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SendMessageToGlobeResponse:
         """
-        ### Usage notes
-        You cannot call the SendMessageToGlobe operation to send messages to the Chinese mainland.
-        ### QPS limit
-        You may call this operation up to 300 times per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
+        @summary Sends a message to regions outside the Chinese mainland.
+        
+        @description ### [](#)Usage notes
+        The SendMessageToGlobe API operation does not support message delivery to the Chinese mainland.
+        ### [](#qps-)QPS limit
+        You can call this operation up to 30 times per second per account. If the number of calls per second exceeds the limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limit when you call this operation.
         
         @param request: SendMessageToGlobeRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -460,6 +490,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.channel_id):
+            query['ChannelId'] = request.channel_id
         if not UtilClient.is_unset(request.from_):
             query['From'] = request.from_
         if not UtilClient.is_unset(request.message):
@@ -494,10 +526,12 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SendMessageToGlobeRequest,
     ) -> dysmsapi_20180501_models.SendMessageToGlobeResponse:
         """
-        ### Usage notes
-        You cannot call the SendMessageToGlobe operation to send messages to the Chinese mainland.
-        ### QPS limit
-        You may call this operation up to 300 times per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
+        @summary Sends a message to regions outside the Chinese mainland.
+        
+        @description ### [](#)Usage notes
+        The SendMessageToGlobe API operation does not support message delivery to the Chinese mainland.
+        ### [](#qps-)QPS limit
+        You can call this operation up to 30 times per second per account. If the number of calls per second exceeds the limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limit when you call this operation.
         
         @param request: SendMessageToGlobeRequest
         @return: SendMessageToGlobeResponse
@@ -510,10 +544,12 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SendMessageToGlobeRequest,
     ) -> dysmsapi_20180501_models.SendMessageToGlobeResponse:
         """
-        ### Usage notes
-        You cannot call the SendMessageToGlobe operation to send messages to the Chinese mainland.
-        ### QPS limit
-        You may call this operation up to 300 times per second. If the number of calls per second exceeds this limit, throttling will be triggered. This can potentially impact your business operations. Therefore, we recommend that you take note of this limit when making calls to this operation.
+        @summary Sends a message to regions outside the Chinese mainland.
+        
+        @description ### [](#)Usage notes
+        The SendMessageToGlobe API operation does not support message delivery to the Chinese mainland.
+        ### [](#qps-)QPS limit
+        You can call this operation up to 30 times per second per account. If the number of calls per second exceeds the limit, throttling is triggered. As a result, your business may be affected. We recommend that you take note of the limit when you call this operation.
         
         @param request: SendMessageToGlobeRequest
         @return: SendMessageToGlobeResponse
@@ -527,7 +563,9 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SendMessageWithTemplateResponse:
         """
-        ### Usage notes
+        @summary Sends a message to the Chinese mainland by using a message template.
+        
+        @description ### Usage notes
         You can call the SendMessageWithTemplate operation to send messages only to the Chinese mainland.
         
         @param request: SendMessageWithTemplateRequest
@@ -536,6 +574,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.channel_id):
+            query['ChannelId'] = request.channel_id
         if not UtilClient.is_unset(request.from_):
             query['From'] = request.from_
         if not UtilClient.is_unset(request.sms_up_extend_code):
@@ -573,7 +613,9 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SendMessageWithTemplateResponse:
         """
-        ### Usage notes
+        @summary Sends a message to the Chinese mainland by using a message template.
+        
+        @description ### Usage notes
         You can call the SendMessageWithTemplate operation to send messages only to the Chinese mainland.
         
         @param request: SendMessageWithTemplateRequest
@@ -582,6 +624,8 @@ class Client(OpenApiClient):
         """
         UtilClient.validate_model(request)
         query = {}
+        if not UtilClient.is_unset(request.channel_id):
+            query['ChannelId'] = request.channel_id
         if not UtilClient.is_unset(request.from_):
             query['From'] = request.from_
         if not UtilClient.is_unset(request.sms_up_extend_code):
@@ -618,7 +662,9 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SendMessageWithTemplateRequest,
     ) -> dysmsapi_20180501_models.SendMessageWithTemplateResponse:
         """
-        ### Usage notes
+        @summary Sends a message to the Chinese mainland by using a message template.
+        
+        @description ### Usage notes
         You can call the SendMessageWithTemplate operation to send messages only to the Chinese mainland.
         
         @param request: SendMessageWithTemplateRequest
@@ -632,7 +678,9 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SendMessageWithTemplateRequest,
     ) -> dysmsapi_20180501_models.SendMessageWithTemplateResponse:
         """
-        ### Usage notes
+        @summary Sends a message to the Chinese mainland by using a message template.
+        
+        @description ### Usage notes
         You can call the SendMessageWithTemplate operation to send messages only to the Chinese mainland.
         
         @param request: SendMessageWithTemplateRequest
@@ -647,13 +695,15 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SmsConversionResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary Delivers one-time password (OTP) message statuses to Alibaba Cloud, which calculates and monitors OTP conversion rates.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         > If you call the SmsConversion operation to query OTP conversion rates, your business may be affected. We recommend that you perform the following operations:
-        > * Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
-        > * Manually degrade your services or use a circuit breaker to automatically degrade services.
+        >  Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
+        >  Manually degrade your services or use a circuit breaker to automatically degrade services.
         
         @param request: SmsConversionRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -692,13 +742,15 @@ class Client(OpenApiClient):
         runtime: util_models.RuntimeOptions,
     ) -> dysmsapi_20180501_models.SmsConversionResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary Delivers one-time password (OTP) message statuses to Alibaba Cloud, which calculates and monitors OTP conversion rates.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         > If you call the SmsConversion operation to query OTP conversion rates, your business may be affected. We recommend that you perform the following operations:
-        > * Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
-        > * Manually degrade your services or use a circuit breaker to automatically degrade services.
+        >  Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
+        >  Manually degrade your services or use a circuit breaker to automatically degrade services.
         
         @param request: SmsConversionRequest
         @param runtime: runtime options for this request RuntimeOptions
@@ -736,13 +788,15 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SmsConversionRequest,
     ) -> dysmsapi_20180501_models.SmsConversionResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary Delivers one-time password (OTP) message statuses to Alibaba Cloud, which calculates and monitors OTP conversion rates.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         > If you call the SmsConversion operation to query OTP conversion rates, your business may be affected. We recommend that you perform the following operations:
-        > * Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
-        > * Manually degrade your services or use a circuit breaker to automatically degrade services.
+        >  Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
+        >  Manually degrade your services or use a circuit breaker to automatically degrade services.
         
         @param request: SmsConversionRequest
         @return: SmsConversionResponse
@@ -755,13 +809,15 @@ class Client(OpenApiClient):
         request: dysmsapi_20180501_models.SmsConversionRequest,
     ) -> dysmsapi_20180501_models.SmsConversionResponse:
         """
-        Metrics:
-        *   Requested OTP messages
-        *   Verified OTP messages
+        @summary Delivers one-time password (OTP) message statuses to Alibaba Cloud, which calculates and monitors OTP conversion rates.
+        
+        @description Metrics:
+        Requested OTP messages
+        Verified OTP messages
         An OTP conversion rate is calculated based on the following formula: OTP conversion rate = Number of verified OTP messages/Number of requested OTP messages.
         > If you call the SmsConversion operation to query OTP conversion rates, your business may be affected. We recommend that you perform the following operations:
-        > * Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
-        > * Manually degrade your services or use a circuit breaker to automatically degrade services.
+        >  Call the SmsConversion operation in an asynchronous manner by configuring queues or events.
+        >  Manually degrade your services or use a circuit breaker to automatically degrade services.
         
         @param request: SmsConversionRequest
         @return: SmsConversionResponse